requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.20
Django==1.9.1
django-extensions==1.6.1
djangorestframework==3.3.2
httplib2==0.9.2
kombu==3.0.33
oauthlib==1.0.3
pytz==2015.7
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.0
six==1.10.0
tweepy==3.5.0
